+++ main.go
@@ -4,11 +4,12 @@
 package main
 
 import (
-	"fmt"
+	"context"
 	"github.com/rs/zerolog"
 	"github.com/steadybit/action-kit/go/action_kit_api/v2"
 	"github.com/steadybit/action-kit/go/action_kit_sdk"
 	"github.com/steadybit/discovery-kit/go/discovery_kit_api"
+	"github.com/steadybit/discovery-kit/go/discovery_kit_sdk"
 	"github.com/steadybit/extension-aws/config"
 	"github.com/steadybit/extension-aws/extaz"
 	"github.com/steadybit/extension-aws/extec2"
@@ -37,9 +38,10 @@ func main() {
 	config.ParseConfiguration()
 	utils.InitializeAwsAccountAccess(config.Config)
 
-	stopCh := make(chan os.Signal, 1)
-	signal.Notify(stopCh, syscall.SIGINT, syscall.SIGTERM, syscall.SIGUSR1)
-	registerHandlers(stopCh)
+	ctx, cancel := SignalCanceledContext()
+	defer cancel()
+
+	registerHandlers(ctx)
 
 	action_kit_sdk.InstallSignalHandler()
 	action_kit_sdk.RegisterCoverageEndpoints()
@@ -50,35 +52,55 @@ func main() {
 	})
 }
 
-func registerHandlers(stopCh chan os.Signal) {
+func SignalCanceledContext() (context.Context, context.CancelFunc) {
+	ctx, cancel := context.WithCancel(context.Background())
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, syscall.SIGINT, syscall.SIGTERM, syscall.SIGUSR1)
+
+	go func() {
+		select {
+		case <-c:
+			cancel()
+		case <-ctx.Done():
+		}
+	}()
+
+	return ctx, func() {
+		signal.Stop(c)
+		cancel()
+	}
+}
+
+func registerHandlers(ctx context.Context) {
 	cfg := config.Config
-	utils.RegisterCommonDiscoveryHandlers()
+	discovery_kit_sdk.Register(utils.NewCommonAttributeDescriber())
+
 	if !cfg.DiscoveryDisabledRds {
-		extrds.RegisterInstanceDiscoveryHandlers(stopCh)
+		discovery_kit_sdk.Register(extrds.NewRdsInstanceDiscovery(ctx))
 		action_kit_sdk.RegisterAction(extrds.NewRdsInstanceRebootAttack())
 		action_kit_sdk.RegisterAction(extrds.NewRdsInstanceStopAttack())
 
-		extrds.RegisterClusterDiscoveryHandlers(stopCh)
+		discovery_kit_sdk.Register(extrds.NewRdsClusterDiscovery(ctx))
 		action_kit_sdk.RegisterAction(extrds.NewRdsClusterFailoverAttack())
 	}
 
 	if !cfg.DiscoveryDisabledZone {
-		extaz.RegisterDiscoveryHandlers(stopCh)
+		discovery_kit_sdk.Register(extaz.NewAzDiscovery(ctx))
 		action_kit_sdk.RegisterAction(extaz.NewAzBlackholeAction())
 	}
 
 	if !cfg.DiscoveryDisabledEc2 {
-		extec2.RegisterDiscoveryHandlers(stopCh)
+		discovery_kit_sdk.Register(extec2.NewEc2InstanceDiscovery(ctx))
 		action_kit_sdk.RegisterAction(extec2.NewEc2InstanceStateAction())
 	}
 
 	if !cfg.DiscoveryDisabledFis {
-		extfis.RegisterFisInstanceDiscoveryHandlers(stopCh)
+		discovery_kit_sdk.Register(extfis.NewFisTemplateDiscovery(ctx))
 		action_kit_sdk.RegisterAction(extfis.NewFisExperimentAction())
 	}
 
 	if !cfg.DiscoveryDisabledLambda {
-		extlambda.RegisterDiscoveryHandlers(stopCh)
+		discovery_kit_sdk.Register(extlambda.NewLambdaDiscovery(ctx))
 		action_kit_sdk.RegisterAction(extlambda.NewInjectStatusCodeAction())
 		action_kit_sdk.RegisterAction(extlambda.NewInjectExceptionAction())
 		action_kit_sdk.RegisterAction(extlambda.NewInjectLatencyAction())
@@ -96,152 +118,7 @@ type ExtensionListResponse struct {
 
 func getExtensionList() ExtensionListResponse {
 	return ExtensionListResponse{
-		ActionList: action_kit_sdk.GetActionList(),
-		DiscoveryList: discovery_kit_api.DiscoveryList{
-			Discoveries:           getDiscoveries(),
-			TargetTypes:           getTargetTypes(),
-			TargetAttributes:      getTargetAttributes(),
-			TargetEnrichmentRules: getEnrichmentRules(),
-		},
+		ActionList:    action_kit_sdk.GetActionList(),
+		DiscoveryList: discovery_kit_sdk.GetDiscoveryList(),
 	}
-}
-
-func getDiscoveries() []discovery_kit_api.DescribingEndpointReference {
-	cfg := config.Config
-	discoveries := make([]discovery_kit_api.DescribingEndpointReference, 0)
-	if !cfg.DiscoveryDisabledRds {
-		discoveries = append(discoveries, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/rds/instance/discovery",
-		})
-		discoveries = append(discoveries, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/rds/cluster/discovery",
-		})
-	}
-	if !cfg.DiscoveryDisabledEc2 {
-		discoveries = append(discoveries, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/ec2/instance/discovery",
-		})
-	}
-	if !cfg.DiscoveryDisabledZone {
-		discoveries = append(discoveries, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/az/discovery",
-		})
-	}
-	if !cfg.DiscoveryDisabledFis {
-		discoveries = append(discoveries, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/fis/template/discovery",
-		})
-	}
-	if !cfg.DiscoveryDisabledLambda {
-		discoveries = append(discoveries, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/lambda/discovery",
-		})
-	}
-	return discoveries
-}
-
-func getTargetTypes() []discovery_kit_api.DescribingEndpointReference {
-	cfg := config.Config
-	targetTypes := make([]discovery_kit_api.DescribingEndpointReference, 0)
-	if !cfg.DiscoveryDisabledRds {
-		targetTypes = append(targetTypes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/rds/instance/discovery/target-description",
-		})
-		targetTypes = append(targetTypes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/rds/cluster/discovery/target-description",
-		})
-	}
-	if !cfg.DiscoveryDisabledEc2 {
-		targetTypes = append(targetTypes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/ec2/instance/discovery/target-description",
-		})
-	}
-	if !cfg.DiscoveryDisabledZone {
-		targetTypes = append(targetTypes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/az/discovery/target-description",
-		})
-	}
-	if !cfg.DiscoveryDisabledFis {
-		targetTypes = append(targetTypes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/fis/template/discovery/target-description",
-		})
-	}
-	if !cfg.DiscoveryDisabledLambda {
-		targetTypes = append(targetTypes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/lambda/discovery/target-description",
-		})
-	}
-	return targetTypes
-}
-
-func getTargetAttributes() []discovery_kit_api.DescribingEndpointReference {
-	cfg := config.Config
-	targetAttributes := make([]discovery_kit_api.DescribingEndpointReference, 0)
-	targetAttributes = append(targetAttributes, discovery_kit_api.DescribingEndpointReference{
-		Method: "GET",
-		Path:   "/common/discovery/attribute-descriptions",
-	})
-	if !cfg.DiscoveryDisabledRds {
-		targetAttributes = append(targetAttributes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/rds/instance/discovery/attribute-descriptions",
-		})
-		targetAttributes = append(targetAttributes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/rds/cluster/discovery/attribute-descriptions",
-		})
-	}
-	if !cfg.DiscoveryDisabledEc2 {
-		targetAttributes = append(targetAttributes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/ec2/instance/discovery/attribute-descriptions",
-		})
-	}
-	if !cfg.DiscoveryDisabledFis {
-		targetAttributes = append(targetAttributes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/fis/template/discovery/attribute-descriptions",
-		})
-	}
-	if !cfg.DiscoveryDisabledLambda {
-		targetAttributes = append(targetAttributes, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/lambda/discovery/attribute-descriptions",
-		})
-	}
-	return targetAttributes
-}
-
-func getEnrichmentRules() []discovery_kit_api.DescribingEndpointReference {
-	cfg := config.Config
-	targetEnrichmentRules := make([]discovery_kit_api.DescribingEndpointReference, 0)
-	if !cfg.DiscoveryDisabledEc2 {
-		targetEnrichmentRules = append(targetEnrichmentRules, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/ec2/instance/discovery/rules/ec2-to-host",
-		})
-		targetEnrichmentRules = append(targetEnrichmentRules, discovery_kit_api.DescribingEndpointReference{
-			Method: "GET",
-			Path:   "/ec2/instance/discovery/rules/ec2-to-kubernetes-node",
-		})
-		for _, targetType := range config.Config.EnrichEc2DataForTargetTypes {
-			targetEnrichmentRules = append(targetEnrichmentRules, discovery_kit_api.DescribingEndpointReference{
-				Method: "GET",
-				Path:   fmt.Sprintf("/ec2/instance/discovery/rules/ec2-to-%s", targetType),
-			})
-		}
-	}
-	return targetEnrichmentRules
 }